// generated/HopL1Bridge/L1_Bridge.ts
export class BonderAdded extends ethereum.Event {
  get params(): BonderAdded__Params {
    return new BonderAdded__Params(this);
  }
}

export class BonderAdded__Params {
  _event: BonderAdded;

  constructor(event: BonderAdded) {
    this._event = event;
  }

  get newBonder(): Address {
    return this._event.parameters[0].value.toAddress();
  }
}

export class BonderRemoved extends ethereum.Event {
  get params(): BonderRemoved__Params {
    return new BonderRemoved__Params(this);
  }
}

export class BonderRemoved__Params {
  _event: BonderRemoved;

  constructor(event: BonderRemoved) {
    this._event = event;
  }

  get previousBonder(): Address {
    return this._event.parameters[0].value.toAddress();
  }
}

export class ChallengeResolved extends ethereum.Event {
  get params(): ChallengeResolved__Params {
    return new ChallengeResolved__Params(this);
  }
}

export class ChallengeResolved__Params {
  _event: ChallengeResolved;

  constructor(event: ChallengeResolved) {
    this._event = event;
  }

  get transferRootId(): Bytes {
    return this._event.parameters[0].value.toBytes();
  }

  get rootHash(): Bytes {
    return this._event.parameters[1].value.toBytes();
  }

  get originalAmount(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }
}

export class MultipleWithdrawalsSettled extends ethereum.Event {
  get params(): MultipleWithdrawalsSettled__Params {
    return new MultipleWithdrawalsSettled__Params(this);
  }
}

export class MultipleWithdrawalsSettled__Params {
  _event: MultipleWithdrawalsSettled;

  constructor(event: MultipleWithdrawalsSettled) {
    this._event = event;
  }

  get bonder(): Address {
    return this._event.parameters[0].value.toAddress();
  }

  get rootHash(): Bytes {
    return this._event.parameters[1].value.toBytes();
  }

  get totalBondsSettled(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }
}

export class Stake extends ethereum.Event {
  get params(): Stake__Params {
    return new Stake__Params(this);
  }
}

export class Stake__Params {
  _event: Stake;

  constructor(event: Stake) {
    this._event = event;
  }

  get account(): Address {
    return this._event.parameters[0].value.toAddress();
  }

  get amount(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }
}

export class TransferBondChallenged extends ethereum.Event {
  get params(): TransferBondChallenged__Params {
    return new TransferBondChallenged__Params(this);
  }
}

export class TransferBondChallenged__Params {
  _event: TransferBondChallenged;

  constructor(event: TransferBondChallenged) {
    this._event = event;
  }

  get transferRootId(): Bytes {
    return this._event.parameters[0].value.toBytes();
  }

  get rootHash(): Bytes {
    return this._event.parameters[1].value.toBytes();
  }

  get originalAmount(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }
}

export class TransferRootBonded extends ethereum.Event {
  get params(): TransferRootBonded__Params {
    return new TransferRootBonded__Params(this);
  }
}

export class TransferRootBonded__Params {
  _event: TransferRootBonded;

  constructor(event: TransferRootBonded) {
    this._event = event;
  }

  get root(): Bytes {
    return this._event.parameters[0].value.toBytes();
  }

  get amount(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }
}

export class TransferRootConfirmed extends ethereum.Event {
  get params(): TransferRootConfirmed__Params {
    return new TransferRootConfirmed__Params(this);
  }
}

export class TransferRootConfirmed__Params {
  _event: TransferRootConfirmed;

  constructor(event: TransferRootConfirmed) {
    this._event = event;
  }

  get originChainId(): BigInt {
    return this._event.parameters[0].value.toBigInt();
  }

  get destinationChainId(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }

  get rootHash(): Bytes {
    return this._event.parameters[2].value.toBytes();
  }

  get totalAmount(): BigInt {
    return this._event.parameters[3].value.toBigInt();
  }
}

export class TransferRootSet extends ethereum.Event {
  get params(): TransferRootSet__Params {
    return new TransferRootSet__Params(this);
  }
}

export class TransferRootSet__Params {
  _event: TransferRootSet;

  constructor(event: TransferRootSet) {
    this._event = event;
  }

  get rootHash(): Bytes {
    return this._event.parameters[0].value.toBytes();
  }

  get totalAmount(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }
}

export class TransferSentToL2 extends ethereum.Event {
  get params(): TransferSentToL2__Params {
    return new TransferSentToL2__Params(this);
  }
}

export class TransferSentToL2__Params {
  _event: TransferSentToL2;

  constructor(event: TransferSentToL2) {
    this._event = event;
  }

  get chainId(): BigInt {
    return this._event.parameters[0].value.toBigInt();
  }

  get recipient(): Address {
    return this._event.parameters[1].value.toAddress();
  }

  get amount(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }

  get amountOutMin(): BigInt {
    return this._event.parameters[3].value.toBigInt();
  }

  get deadline(): BigInt {
    return this._event.parameters[4].value.toBigInt();
  }

  get relayer(): Address {
    return this._event.parameters[5].value.toAddress();
  }

  get relayerFee(): BigInt {
    return this._event.parameters[6].value.toBigInt();
  }
}

export class Unstake extends ethereum.Event {
  get params(): Unstake__Params {
    return new Unstake__Params(this);
  }
}

export class Unstake__Params {
  _event: Unstake;

  constructor(event: Unstake) {
    this._event = event;
  }

  get account(): Address {
    return this._event.parameters[0].value.toAddress();
  }

  get amount(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }
}

export class WithdrawalBondSettled extends ethereum.Event {
  get params(): WithdrawalBondSettled__Params {
    return new WithdrawalBondSettled__Params(this);
  }
}

export class WithdrawalBondSettled__Params {
  _event: WithdrawalBondSettled;

  constructor(event: WithdrawalBondSettled) {
    this._event = event;
  }

  get bonder(): Address {
    return this._event.parameters[0].value.toAddress();
  }

  get transferId(): Bytes {
    return this._event.parameters[1].value.toBytes();
  }

  get rootHash(): Bytes {
    return this._event.parameters[2].value.toBytes();
  }
}

export class WithdrawalBonded extends ethereum.Event {
  get params(): WithdrawalBonded__Params {
    return new WithdrawalBonded__Params(this);
  }
}

export class WithdrawalBonded__Params {
  _event: WithdrawalBonded;

  constructor(event: WithdrawalBonded) {
    this._event = event;
  }

  get transferId(): Bytes {
    return this._event.parameters[0].value.toBytes();
  }

  get amount(): BigInt {
    return this._event.parameters[1].value.toBigInt();
  }
}

export class Withdrew extends ethereum.Event {
  get params(): Withdrew__Params {
    return new Withdrew__Params(this);
  }
}

export class Withdrew__Params {
  _event: Withdrew;

  constructor(event: Withdrew) {
    this._event = event;
  }

  get transferId(): Bytes {
    return this._event.parameters[0].value.toBytes();
  }

  get recipient(): Address {
    return this._event.parameters[1].value.toAddress();
  }

  get amount(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }

  get transferNonce(): Bytes {
    return this._event.parameters[3].value.toBytes();
  }
}